moved force theme to parent screen
