+++ android/src/main/java/io/parity/signer/screens/initial/explanation/OnboardingScreenFull.kt
@@ -1,5 +1,6 @@
 package io.parity.signer.screens.initial.explanation
 
+import androidx.appcompat.app.AppCompatDelegate
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.captionBarPadding
@@ -7,6 +8,7 @@ import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
 import androidx.compose.ui.Modifier
 import io.parity.signer.domain.Callback
 
@@ -20,6 +22,7 @@ fun OnboardingExplanationScreenFull(navigateNext: Callback) {
 			.captionBarPadding()
 			.statusBarsPadding()
 	) {
+		ForceDarkTheme()
 		HorizontalPager(pageCount = 4) {
 			when (it) {
 				0 -> OnboardingScreen1(navigateNext)
@@ -27,6 +30,17 @@ fun OnboardingExplanationScreenFull(navigateNext: Callback) {
 				2 -> OnboardingScreen3(navigateNext)
 				3 -> OnboardingScreen4(navigateNext)
 			}
+		}
+	}
+}
+
+@Composable
+private fun ForceDarkTheme() {
+	DisposableEffect(key1 = Unit) {
+		val original = AppCompatDelegate.getDefaultNightMode()
+		AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
+		onDispose {
+			AppCompatDelegate.setDefaultNightMode(original)
 		}
 	}
 }
